Separate the VM controller and state driver into distinct modules
Move code around so that the VM controller and the state driver are in separate
modules, in part to keep the VM controller module from getting too large and in
part to start setting up a visibility barrier between the two components.
